Add user-to-user matching feature and related APIs
- Implement matching logic to find similar users based on emotion entries
- Add endpoints for creating matchings, retrieving candidates, and creating rooms from matchings
- Update IMatchingService and MatchingService for user-specific access and ownership checks
- Introduce CreateMatchingRequest DTO
- Add EmotionController and MatchingController with authentication and access control
- Refine similarity score precision in MatchingRepository
- Remove UserPresences DbSet from ApplicationDbContext
- Minor cleanup in CommunityController's Leave method

+++ Infracstructure/Repositories/MatchingRepository.cs
@@ -58,7 +58,7 @@ public class MatchingRepository : IMatchingRepository
             ranked_candidates AS (
                 SELECT DISTINCT ON (candidate_entry.user_id)
                     candidate_entry.user_id AS ""CandidateUserId"",
-                    CAST(1 - (source_embedding.embedding <=> candidate_embedding.embedding) AS double precision) AS ""SimilarityScore"",
+                    CAST(1 - (source_embedding.embedding <=> candidate_embedding.embedding) AS numeric(6,5)) AS ""SimilarityScore"",
                     CASE
                         WHEN candidate_entry.community_id IS NOT DISTINCT FROM source_entry.community_id
                             THEN 'Same community and similar emotional context.'
